+++ packages/network/src/context.ts
@@ -19,10 +19,10 @@ export interface WalletContext {
 }
 
 export const sodiumContext: WalletContext = {
-  entryPointAddress: "0x0aE1B76389397Dc81c16eB8e2dEb0C592D3C873c",
-  singletonAddress: "0x09f30e8aC8F1bb14858477b72643eDa238d24339",
+  entryPointAddress: "0xDCa3e88DE9f6A2d55917d1DFE0C81aEDD363c0f9",
+  singletonAddress: "0x5754BDe51007daC3CADD085E6047e38E30249D1B",
   genesisSingletonAddress: "0x09f30e8aC8F1bb14858477b72643eDa238d24339",
-  defaultHandlerAddress: "0xb961f9F277386e449f324A9B8A0b3FDE837BbF08",
+  defaultHandlerAddress: "0x8436173e1D2599Ed3d0aF23C5e231cAee39d67c9",
   walletCreatorAddress: "0xCA25952469BBae82c63A7A75D8aBc7dBf9442e18",
 
   modules: {