+++ examples/my-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # my-app
 
+## 1.9.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/provider@1.10.0
+
 ## 1.9.0
 
 ### Minor Changes