+++ packages/config/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @0xsodium/config
 
+## 1.10.0
+
+### Minor Changes
+
+- support temp upgrade
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/multicall@1.10.0
+  - @0xsodium/network@1.10.0
+  - @0xsodium/utils@1.10.0
+
 ## 1.9.0
 
 ### Minor Changes