+++ packages/graphquery/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @0xsodium/graphquery
 
+## 1.10.0
+
+### Minor Changes
+
+- support temp upgrade
+
 ## 1.9.0
 
 ### Minor Changes