+++ packages/sdk4337/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @0xsodium/sdk4337
 
+## 1.10.0
+
+### Minor Changes
+
+- support temp upgrade
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/config@1.10.0
+  - @0xsodium/estimator@1.10.0
+  - @0xsodium/network@1.10.0
+  - @0xsodium/transactions@1.10.0
+  - @0xsodium/utils@1.10.0
+
 ## 1.9.0
 
 ### Minor Changes