+++ packages/simulator/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @0xsodium/simulator
 
+## 1.10.0
+
+### Minor Changes
+
+- support temp upgrade
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/transactions@1.10.0
+
 ## 1.9.0
 
 ### Minor Changes